pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pr-review-comments-summary"
version = "0.1.0"
description = "PR review comments summarization tool using Azure OpenAI"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "requests",
    "streamlit",
    "openai",
    "chromadb",
    "sentence-transformers",
    "torchvision",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "tiktoken",
    "watchdog>=6.0.0",
    "fpdf2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.29.0",
    "langfuse>=2.0,<3.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "black>=23.3.0",
    "flake8>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E501"]
